Feat: 필드 전역 상태 변수 초기화 로직 추가가
