+++ src/hooks/apis/auth/useCheckNicknameDuplication.ts
@@ -1,5 +1,5 @@
 import { checkNicknameDuplication } from '@/api/authService';
-import { CheckVerifyFieldProps } from '@/types/propsTypes';
+import { CheckVerifyFieldProps } from '@/types/fieldType';
 import { useMutation } from '@tanstack/react-query';
 
 const useCheckNicknameDuplication = ({ setError, setFieldStatus }: CheckVerifyFieldProps) => {
