+++ src/hooks/apis/auth/useCheckUsernameDuplication.ts
@@ -1,5 +1,5 @@
 import { checkUsernameDuplication } from '@/api/authService';
-import { CheckVerifyFieldProps } from '@/types/propsTypes';
+import { CheckVerifyFieldProps } from '@/types/fieldType';
 import { useMutation } from '@tanstack/react-query';
 
 const useCheckUsernameDuplication = ({ setError, setFieldStatus }: CheckVerifyFieldProps) => {
